Move api scripts to its package
