+++ package.json
@@ -1,19 +1,19 @@
 {
   "name": "forklift-ui",
   "version": "0.0.1",
-  "main": "packages/api/server.js",
+  "main": "packages/api/src/server.js",
   "license": "Apache-2.0",
   "private": true,
   "scripts": {
     "dr:surge": "node dr-surge.js",
-    "setup:dev:remote": "node --trace-deprecation ./scripts/remote-dev.js",
+    "setup:dev:remote": "node ./packages/api/scripts/remote-dev.js",
     "build": "npm run clean && NODE_ENV=production $(npm bin)/webpack --config ./config/webpack.prod.js",
     "build:remote": "DATA_SOURCE=remote npm run build",
     "build:mock": "DATA_SOURCE=mock FORKLIFT_UI_GIT_COMMIT=$(git rev-parse --short HEAD) npm run build && npm run dr:surge",
-    "start": "node --trace-deprecation ./packages/api/src/server.js",
+    "start": "node ./packages/api/src/server.js",
     "start:remote": "DATA_SOURCE=remote npm run start",
     "start:mock": "DATA_SOURCE=mock npm run start",
-    "start:dev": "./scripts/start-dev.sh",
+    "start:dev": "./packages/api/scripts/start-dev.sh",
     "start:dev:remote": "npm run setup:dev:remote && DATA_SOURCE=remote npm run start:dev",
     "start:dev:mock": "DATA_SOURCE=mock npm run start:dev",
     "format": "npm run format -w packages/web && npm run format -w packages/qe-tests",
