+++ packages/web/package.json
@@ -40,20 +40,20 @@
     "@patternfly/react-table": "^4.44.4",
     "@patternfly/react-tokens": "^4.27.4",
     "axios": "^0.21.2",
-    "dayjs-ext": "^2.2.0",
     "dayjs": "^1.10.5",
+    "dayjs-ext": "^2.2.0",
     "fast-deep-equal": "^3.1.3",
     "file-saver": "^2.0.5",
     "js-yaml": "^4.1.0",
     "netmask": "^2.0.2",
     "node-fetch": "^2.6.1",
     "q": "^1.5.1",
+    "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-query": "^3.17.0",
     "react-router-dom": "^5.2.0",
     "react-router-last-location": "^2.0.1",
     "react-syntax-highlighter": "^15.4.3",
-    "react": "^17.0.2",
     "yup": "^0.32.9"
   }
 }
